Refactor to only use Int for id

+++ app/src/main/java/de/janhopp/gymbro/model/execution/Set.kt
@@ -1,25 +1,25 @@
 package de.janhopp.gymbro.model.execution
 
 interface Set {
-    val id: Long
+    val id: Int
     val index: Int
 }
 
 data class WeightSet(
-    override val id: Long,
+    override val id: Int,
     override val index: Int,
     val weight: Int,
     val reps: Int,
 ) : Set
 
 data class BodyWeightSet(
-    override val id: Long,
+    override val id: Int,
     override val index: Int,
     val reps: Int,
 ) : Set
 
 data class CardioSet(
-    override val id: Long,
+    override val id: Int,
     override val index: Int,
     val duration: Int,
 ) : Set
